Template /etc/assisted/rendezvous-host.env
The template for rendezvous-host.env will have .RendezvousIP as
field name to represent the Rendezvous IP address.
